fix zora nfts
