+++ components/common/CharmEditor/components/nft/NFTNodeView.tsx
@@ -19,9 +19,9 @@ import charmClient from 'charmClient';
 import { Button } from 'components/common/Button';
 import { InputSearchBlockchain } from 'components/common/form/InputSearchBlockchain';
 import LoadingComponent from 'components/common/LoadingComponent';
-import { supportedMainnets as supportedMainnetsByAlchemy } from 'lib/blockchain/provider/alchemy';
-import { supportedMainnets as supportedMainnetsByAnkr } from 'lib/blockchain/provider/ankr';
-import { supportedNetworks as supportedNetworksByZora } from 'lib/blockchain/provider/zora/zoraClient';
+import { supportedMainnets as supportedMainnetsByAlchemy } from 'lib/blockchain/provider/alchemy/client';
+import { supportedMainnets as supportedMainnetsByAnkr } from 'lib/blockchain/provider/ankr/config';
+import { supportedNetworks as supportedNetworksByZora } from 'lib/blockchain/provider/zora/config';
 import { MIN_IMAGE_WIDTH } from 'lib/prosemirror/plugins/image/constants';
 
 import { enableDragAndDrop } from '../../utils';
