+++ lib/blockchain/provider/ankr/client.ts
@@ -9,35 +9,14 @@ import { ethers } from 'ethers';
 
 import { getNFTUrl } from 'components/common/CharmEditor/components/nft/utils';
 import { paginatedCall } from 'lib/utilities/async';
-import { typedKeys } from 'lib/utilities/objects';
 
-import type { NFTData } from '../getNFTs';
+import type { NFTData } from '../../getNFTs';
+
+import type { SupportedChainId } from './config';
+import { advancedAPIEndpoint, ankrAdvancedApis, rpcApis } from './config';
 
 // 50 requests/minute for Public tier - https://www.ankr.com/docs/rpc-service/service-plans/#rate-limits
 const rateLimiter = RateLimit(0.8);
-
-// Find supported chains:  https://www.npmjs.com/package/@ankr.com/ankr.js
-// Note: we commented out chains already supported by alchemy
-const ankrAdvancedApis = {
-  // 1: 'eth',
-  // 5: 'eth_goerli',
-  // 10: 'optimism',
-  56: 'bsc',
-  // 137: 'polygon',
-  250: 'fantom',
-  // 42161: 'arbitrum',
-  43114: 'avalanche',
-  5000: 'mantle'
-} as const;
-// https://docs.alchemy.com/docs/why-use-alchemy#-blockchains-supported
-export const supportedChainIds = typedKeys(ankrAdvancedApis);
-export type SupportedChainId = (typeof supportedChainIds)[number];
-
-// We can still support chains that dont have an advanced api
-const rpcApis: SupportedChainId[] = [5000];
-export const supportedMainnets: SupportedChainId[] = [56, 250, 43114, 5000];
-
-const advancedAPIEndpoint = `https://rpc.ankr.com/multichain/${process.env.ANKR_API_ID}`;
 
 function getRPCEndpoint(chainId: SupportedChainId) {
   const chainPath = ankrAdvancedApis[chainId];
