+++ lib/blockchain/provider/zora/config.ts
@@ -1,0 +1,4 @@
+import { zora, zoraTestnet } from 'viem/chains';
+
+export const supportedNetworks = [zora.id, zoraTestnet.id];
+export type SupportedChainId = (typeof supportedNetworks)[number];
